feat: sJson integration, TOCTOU fixes, multi-thread resume, locale translations
- Replace json_builder_t with sJson tree builder in handlers_dump.c
  (arena-based, type-safe JSON with hex_u64/hex_preview helpers)
- Add sjson.h wrapper, sjson_impl.c anchor, external/sjson/sJson.c v1.1.0
- Fix TOCTOU race: replace is_attached()/detach()/attach() patterns
  with memdbg_debugger_conditional_attach() in handlers_dump,
  handlers_process, and handlers_tracer
- Fix multi-thread resume: continue() now iterates all threads on INT3
  instead of stopping at the first one
- Add unit tests:
  - test_conditional_attach (same PID, different PID, attach error)
  - test_multithread_continue (two threads on INT3, verify both stepped)
  - test_detach_idempotency (never attached, double-detach, 5x detach)
  - test_sjson.c (68 tests: arena lifecycle, primitives, hex, objects,
    arrays, nested dump structure, measure/write errors, empty structures)
  - test_legacy_process_e2e.c (legacy compat process/memory E2E)
- Add 23 locale keys (process tree, JSON dump, syscall viewer) to all
  9 languages with real translations (de, es, fr, ja, ko, pt, ru)
- Update Makefile with test-sjson and test-legacy-process-e2e targets
- Keep check-locales and locale manifest in sync (1062 keys across 9 locales)

+++ src/core/daemon/handlers_scan.c
@@ -24,7 +24,7 @@ static memdbg_status_t send_scan_result(int fd, const memdbg_packet_header_t *re
   if (prefix_size < MEMDBG_PROTOCOL_MAX_PACKET) {
     max_count = (MEMDBG_PROTOCOL_MAX_PACKET - prefix_size) / entry_size;
   }
-  uint32_t send_count = result->count;
+  uint32_t send_count = (uint32_t)result->count;
   bool truncated = result->truncated ? true : false;
   if ((size_t)send_count > max_count) {
     send_count = (uint32_t)max_count;
